+++ pyproject.toml
@@ -13,10 +13,9 @@ indexify-extractor = "indexify_extractor_sdk.main:typer_app"
 python = "^3.10"
 typer = {extras = ["all"], version = "^0.9.0"}
 pydantic = "^2.6.1"
-pydantic-settings = "^2.1.0"
 langchain = "^0.1.5"
 transformers = "^4.37.2"
-# torch = "^2.2.0"
+torch = "^2.2.0"
 fastapi = "^0.109.2"
 nanoid = "^2.0.0"
 grpcio = "^1.60.1"
